Log invalid permissions in AuthSubject not reject
By rejecting any invalid permissions when creating an AuthSubject, it
means we can *never* remove a permission from the system, because then
there is an error whenever we rehydrate an older event's auth subject
that has that permission. We log them now at a debug level for
visibility, but otherwise let it stand.